Remove PinInput from input component and use as independent component
